@@ -4,6 +4,7 @@
   "description": "Real-time client/server communication over UDP using WebRTC and Node.js",
   "main": "./lib/index.js",
   "types": "./lib/index.d.ts",
+  "type": "module",
   "author": "Yannick Deubel (https://github.com/yandeu)",
   "license": "BSD-3-Clause",
   "keywords": [
@@ -35,7 +36,7 @@
     "lib": "lib"
   },
   "engines": {
-    "node": ">=12"
+    "node": "^14.15 || >=16"
   },
   "scripts": {
     "build": "tsc --build tsconfig.json",
