@@ -1,3 +1,3 @@
-import client from './geckos/channel'
+import client from './geckos/channel.js'
 
 export default client
