@@ -1,5 +1,5 @@
-import { ChannelId, Data, RawMessage } from '@geckos.io/common/lib/types'
-import client, { ClientChannel } from './geckos/channel'
+import { ChannelId, Data, RawMessage } from '@geckos.io/common/lib/types.js'
+import client, { ClientChannel } from './geckos/channel.js'
 
 export default client
 export { client as geckos }
