@@ -7,7 +7,7 @@ const TerserPlugin = require('terser-webpack-plugin')
 module.exports = env => {
   return {
     mode: 'production',
-    entry: path.resolve(__dirname, '../src/bundle.ts'),
+    entry: path.resolve(__dirname, '../lib/bundle.js'),
     output: {
       path: path.resolve(__dirname, `${env.path}`),
       filename: `geckos.io-client.${env.packageVersion}.min.js`,
