@@ -1,5 +1,5 @@
-import makeRandomId from './makeRandomId'
-import runInterval from './runInterval'
+import makeRandomId from './makeRandomId.js'
+import runInterval from './runInterval.js'
 
 const makeReliable = (options: any, cb: Function) => {
   const { interval = 150, runs = 10 } = options
