@@ -1,6 +1,6 @@
-import { ChannelId, Data, RawMessage } from '@geckos.io/common/lib/types'
-import server, { GeckosServer, ServerChannel } from './geckos/server'
-import iceServers from '@geckos.io/common/lib/iceServers'
+import { ChannelId, Data, RawMessage } from '@geckos.io/common/lib/types.js'
+import server, { GeckosServer, ServerChannel } from './geckos/server.js'
+import iceServers from '@geckos.io/common/lib/iceServers.js'
 
 export default server
 export { server as geckos, iceServers }
